% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_by_month.R
\name{download_by_month}
\alias{download_by_month}
\title{Get the monthly average download counts}
\usage{
download_by_month(vec)
}
\arguments{
\item{vec}{A vector of character values. Refer the relavant information about a package you are looking for.}
}
\value{
\code{data.frame}. A data.frame with the average monthly download counts of these packages.
}
\description{
Get the monthly average download counts
}
\seealso{
download_trend
}
\author{
Jiaxiang Li \email{alex.lijiaxiang@foxmail.com}
}
