% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_log.R
\name{download_log}
\alias{download_log}
\title{Get these packages' download counts}
\usage{
download_log(vec)
}
\arguments{
\item{vec}{A vector of character values. Refer the relavant information about a package you are looking for.}
}
\value{
\code{data.frame}. A data.frame with monthly download counts of these packages.
}
\description{
Get these packages' download counts
}
\seealso{
download_by_month
}
\author{
Jiaxiang Li \email{alex.lijiaxiang@foxmail.com}
}
