% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_trend.R
\name{download_trend}
\alias{download_trend}
\title{Get the trend of the download counts}
\usage{
download_trend(vec)
}
\arguments{
\item{vec}{A vector of character values. Refer the relavant information about a package you are looking for.}
}
\value{
\code{ggplot2}.
}
\description{
Get the trend of the download counts
}
\seealso{
download_by_month
}
\author{
Jiaxiang Li \email{alex.lijiaxiang@foxmail.com}
}
