% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_list.R
\name{pkg_list}
\alias{pkg_list}
\title{Get the list of packages}
\usage{
pkg_list(vec)
}
\arguments{
\item{vec}{A vector of character values. Refer the relavant information about a package you are looking for.}
}
\value{
tibble. A tibble with a list of packages.
}
\description{
Get the list of packages
}
\seealso{
download_log
}
\author{
Jiaxiang Li \email{alex.lijiaxiang@foxmail.com}
}
