% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repex_in.R
\name{regex_in}
\alias{regex_in}
\title{Put regular expression on a string and output T/F}
\usage{
regex_in(x, regex = ".")
}
\arguments{
\item{x}{Character. A column of character.}

\item{regex}{Character. The regular expression rule defined.}
}
\value{
Logical. A column of logical values.
}
\description{
Put regular expression on a string and output T/F
}
\seealso{
str_subset
}
\author{
Jiaxiang Li \email{alex.lijiaxiang@foxmail.com}
}
